requirements.txt
pandas
pydantic
requests
python-dotenv
google-cloud-storage
google-cloud-bigquery
matplotlib
statsmodels
pyarrow